feat(niivue): add a linear vs cubic B-spline interpolation example
Pull the filter A/B tooling out of the chunk-streaming demo into its own
page. Two audiences: debugging a render whose reconstruction filter is
under suspicion, and producing higher-quality still figures, where the
2x fine-march cost of the cubic filter does not matter.

vox.interp.html loads a standard volume in the render view and carries
the Interp select, the split/flip A/B comparison with a header divider
slider, sample-rate and illumination sliders, a ms/frame benchmark over
both filters, and a PNG export (the split composite when one is up).

range.html keeps the plain Interp select, since a chunked stream is
where halo correctness matters, and drops the A/B machinery.
