+++ packages/niivue/examples/index.html
@@ -141,6 +141,7 @@
           <li><a href="./vox.modulate.afni.html" target="_blank" rel="noopener noreferrer">Modulate DTI V1 (AFNI)</a></li>
           <li><a href="./vox.modulate.scalar.html" target="_blank" rel="noopener noreferrer">Modulate scalar overlay by mask</a></li>
           <li><a href="./vox.window.html" target="_blank" rel="noopener noreferrer">Intensity window (min/max sliders)</a></li>
+          <li><a href="./vox.interp.html" target="_blank" rel="noopener noreferrer">Interpolation A/B (linear vs cubic B-spline)</a></li>
           <li><a href="./vox.draw.html" target="_blank" rel="noopener noreferrer">Drawing</a></li>
           <li><a href="./vox.draw.explode.html" target="_blank" rel="noopener noreferrer">Drawing on exploded blocks (2D + 3D)</a></li>
           <li><a href="./vox.tiled.html" target="_blank" rel="noopener noreferrer">Tiled (oversized) volumes</a></li>
